Hoist medium-neutral view semantics out of the terminal renderer
src/semantics owns what every renderer must agree on: the zoom ladder
(ZoomStep, clampZoom, zoomLabel, and zoomMode — the one home for WHICH
steps switch mode), the neutral-kind rule, aggregateMap (group collapse
for the far zoom) and flipForSequence (sequence time orientation). The
terminal renderer consumes them and re-exports the ladder so terminal
consumers keep one import site; zoomGeometry now derives its mode from
zoomMode and owns only whitespace numbers. No behavior change: all 177
tests including render snapshots pass unmodified.
